Styles revision (icons background, layout in sm/md screens, paddings)
