+++ apps/web/app/[lang]/components/howitworks/HowItWorks.types.ts
