+++ apps/web/app/[lang]/components/howitworks/helpers.tsx
