+++ apps/web/app/[lang]/components/howitworks/icons/ApplyIcon.tsx
