+++ apps/web/app/[lang]/components/howitworks/icons/SearchIcon.tsx
