+++ apps/web/app/[lang]/components/howitworks/icons/TakePartIcon.tsx
