+++ apps/web/app/[lang]/components/howitworks/icons/index.ts
