+++ apps/web/app/[lang]/components/howitworks/index.ts
